refactor: Move cozy helper functions to standalone files
